add umap emb before uploading
