+++ dataquality/loggers/base_logger.py
@@ -17,9 +17,10 @@ from dataquality.loggers.logger_config.base_logger_config import (
 from dataquality.schemas.split import Split, conform_split
 from dataquality.schemas.task_type import TaskType
 from dataquality.utils.cloud import is_galileo_cloud
-from dataquality.utils.dq_logger import _shutil_rmtree_retry, upload_dq_log_file
+from dataquality.utils.dq_logger import upload_dq_log_file
 from dataquality.utils.imports import hf_available, tf_available, torch_available
 from dataquality.utils.tf import is_tf_2
+from dataquality.utils.file import _shutil_rmtree_retry
 
 T = TypeVar("T", bound="BaseGalileoLogger")
